+++ api/lib/application/target-profiles/target-profile-controller.js
@@ -1,9 +1,9 @@
 import { usecases as devcompUsecases } from '../../../src/devcomp/domain/usecases/index.js';
 import * as trainingSummarySerializer from '../../../src/devcomp/infrastructure/serializers/jsonapi/training-summary-serializer.js';
+import * as targetProfileForAdminSerializer from '../../../src/prescription/target-profile/infrastructure/serializers/jsonapi/target-profile-for-admin-serializer.js';
 import * as targetProfileSerializer from '../../../src/prescription/target-profile/infrastructure/serializers/jsonapi/target-profile-serializer.js';
 import { usecases } from '../../domain/usecases/index.js';
 import { DomainTransaction } from '../../infrastructure/DomainTransaction.js';
-import * as targetProfileForAdminSerializer from '../../infrastructure/serializers/jsonapi/target-profile-for-admin-serializer.js';
 
 const getTargetProfileForAdmin = async function (request, h, dependencies = { targetProfileForAdminSerializer }) {
   const targetProfileId = request.params.id;
