+++ api/tests/prescription/target-profile/unit/infrastructure/serializers/jsonapi/target-profile-for-admin-serializer_test.js
@@ -1,6 +1,6 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/target-profile-for-admin-serializer.js';
-import { TargetProfileForAdmin } from '../../../../../src/prescription/target-profile/domain/models/TargetProfileForAdmin.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import { TargetProfileForAdmin } from '../../../../../../../src/prescription/target-profile/domain/models/TargetProfileForAdmin.js';
+import * as serializer from '../../../../../../../src/prescription/target-profile/infrastructure/serializers/jsonapi/target-profile-for-admin-serializer.js';
+import { domainBuilder, expect } from '../../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | target-profile-for-admin-serializer', function () {
   describe('#serialize', function () {
